hide tooltip on mobile
