+++ src/components/balances/HiddenTokenButton/index.tsx
@@ -7,6 +7,11 @@ import { VisibilityOutlined } from '@mui/icons-material'
 import Track from '@/components/common/Track'
 
 import css from './styles.module.css'
+import useABTesting from '@/services/tracking/useABTesting'
+import { AbTest } from '@/services/tracking/abTesting'
+import { OnboardingTooltip } from '@/components/common/OnboardingTooltip'
+
+const LS_ONBOARDING = 'ONBOARDING_HIDDEN_TOKEN_BUTTON'
 
 const HiddenTokenButton = ({
   toggleShowHiddenAssets,
@@ -17,28 +22,43 @@ const HiddenTokenButton = ({
 }): ReactElement | null => {
   const { balances } = useBalances()
   const currentHiddenAssets = useHiddenTokens()
+  const isTooltipShown = useABTesting(AbTest.HIDE_TOKEN_PROMO)
 
   const hiddenAssetCount =
     balances.items?.filter((item) => currentHiddenAssets.includes(item.tokenInfo.address)).length || 0
 
   return (
-    <Track {...ASSETS_EVENTS.SHOW_HIDDEN_ASSETS}>
-      <Button
-        className={css.hiddenTokenButton}
-        sx={{ gap: 1, padding: 1, borderWidth: '1px !important', borderColor: ({ palette }) => palette.border.main }}
-        disabled={showHiddenAssets}
-        onClick={toggleShowHiddenAssets}
-        data-testid="toggle-hidden-assets"
-        variant="outlined"
-      >
-        <>
-          <VisibilityOutlined fontSize="small" />
-          <Typography fontSize="small">
-            {hiddenAssetCount === 0 ? 'Hide tokens' : `${hiddenAssetCount} Hidden token(s)`}{' '}
-          </Typography>
-        </>
-      </Button>
-    </Track>
+    <OnboardingTooltip
+      className={css.hiddenTokenButton}
+      initiallyShown={isTooltipShown}
+      widgetLocalStorageId={LS_ONBOARDING}
+      text="Spam or unwanted tokens in your asset list? Hide them now!"
+    >
+      <div>
+        <Track {...ASSETS_EVENTS.SHOW_HIDDEN_ASSETS}>
+          <Button
+            className={css.hiddenTokenButton}
+            sx={{
+              gap: 1,
+              padding: 1,
+              borderWidth: '1px !important',
+              borderColor: ({ palette }) => palette.border.main,
+            }}
+            disabled={showHiddenAssets}
+            onClick={toggleShowHiddenAssets}
+            data-testid="toggle-hidden-assets"
+            variant="outlined"
+          >
+            <>
+              <VisibilityOutlined fontSize="small" />
+              <Typography fontSize="small">
+                {hiddenAssetCount === 0 ? 'Hide tokens' : `${hiddenAssetCount} Hidden token(s)`}{' '}
+              </Typography>
+            </>
+          </Button>
+        </Track>
+      </div>
+    </OnboardingTooltip>
   )
 }
