+++ src/pages/balances/index.tsx
@@ -11,17 +11,11 @@ import PagePlaceholder from '@/components/common/PagePlaceholder'
 import NoAssetsIcon from '@/public/images/balances/no-assets.svg'
 import HiddenTokenButton from '@/components/balances/HiddenTokenButton'
 import CurrencySelect from '@/components/balances/CurrencySelect'
-import { OnboardingTooltip } from '@/components/common/OnboardingTooltip'
-import { AbTest } from '@/services/tracking/abTesting'
-import useABTesting from '@/services/tracking/useABTesting'
-
-const LS_ONBOARDING = 'ONBOARDING_HIDDEN_TOKEN_BUTTON'
 
 const Balances: NextPage = () => {
   const { loading, error } = useBalances()
   const [showHiddenAssets, setShowHiddenAssets] = useState(false)
   const toggleShowHiddenAssets = () => setShowHiddenAssets((prev) => !prev)
-  const isTooltipShown = useABTesting(AbTest.HIDE_TOKEN_PROMO)
 
   return (
     <>
@@ -31,20 +25,7 @@ const Balances: NextPage = () => {
 
       <AssetsHeader>
         <Box display="flex" flexDirection="row" alignItems="center" gap={1}>
-          {
-            <OnboardingTooltip
-              initiallyShown={isTooltipShown}
-              widgetLocalStorageId={LS_ONBOARDING}
-              text="Spam or unwanted tokens in your asset list? Hide them now!"
-            >
-              <div>
-                <HiddenTokenButton
-                  showHiddenAssets={showHiddenAssets}
-                  toggleShowHiddenAssets={toggleShowHiddenAssets}
-                />
-              </div>
-            </OnboardingTooltip>
-          }
+          <HiddenTokenButton showHiddenAssets={showHiddenAssets} toggleShowHiddenAssets={toggleShowHiddenAssets} />
           <CurrencySelect />
         </Box>
       </AssetsHeader>
